kindrio/poc/requirements.txt
penai==0.27.2
langchain==0.0.177